Create course contents
This course introduces you to the fundamental theory of computational fluid dynamics with a basic hands on experience of the CFD tools. The course will kick start your CFD journey and will give you flavor of all basic modalities of the CFD. The course is designed in such a way that a coherent linkage is established between a ranges of CFD concepts and practical learning. The course is divided into four modules. The first three modules are focused towards the theoretical understanding of the fundamentals of CFD. The last module is focused towards the practical hands on experience of the CFD tools which familiarize the user with the CFD software.